+++ docs/source/examples/patterns.py
@@ -26,4 +26,3 @@ try:
 except KeyboardInterrupt:
     observer.stop()
 observer.join()
-